fix: share guarded Classic launch flow
