+++ src/main/helpers/index.ts
@@ -96,6 +96,7 @@ export const getThemeSoundPath = (
 export * from "./reg-parser";
 export * from "./launch-game";
 export * from "./launch-classics-game";
+export * from "./open-classics-game";
 export * from "./download-error-handler";
 export * from "./download-game-helper";
 export * from "./platform-to-system";
